requirements.txt
ultralytics==8.3.158
streamlit==1.37.1
numpy==1.26.4
opencv-python==4.10.0.84
requests==2.32.3
Pillow==10.4.0
pandas==2.2.2
matplotlib==3.8.4
